Add: noto-sans-font example

+++ src/screen.rs
@@ -1,6 +1,6 @@
 pub mod interface {
     pub trait Write {
-        fn draw(&self, x: usize, y: usize, c: usize);
+        fn draw(&self, x: usize, y: usize, c: u32);
     }
 
     pub trait All = Write;
